requirements.txt
# Kalpataru - Agricultural AI Platform

## Core Dependencies
flask==3.0.0
streamlit==1.29.0
pydantic==2.5.3

## Data Processing
numpy==1.26.2
pandas==2.1.4
Pillow==10.1.0

## Machine Learning & Deep Learning
tensorflow>=2.15.0
xgboost==2.0.3
prophet==1.1.5
scikit-learn==1.3.2

## Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

## API & Web
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0

## Utilities
python-dateutil==2.8.2
six==1.16.0

## Logging
colorlog==6.8.0

## Development
pytest==7.4.3
black==23.12.1
flake8==6.1.0
